requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.14.0
numpy==1.24.3
pandas==2.1.1
pillow==10.0.1
plotly==5.18.0
scikit-learn==1.3.2
pytest==8.3.4
pytest-cov==6.0.0
flake8==6.1.0
black==23.11.0